Lift remaining contexts
